+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix app-admission-controller webhook name in validation error matchers.
+
 ## [5.2.2] - 2021-08-17
 
 ### Fixed